mysql_secure_installation_complete

+++ chef-ispconfig3/cookbooks/ispconfig3/recipes/postfix.rb
@@ -39,32 +39,29 @@ template '/etc/mysql/mariadb.conf.d/mysqld.cnf' do
     source 'mysqld.cnf.erb'
 end
 
-#--------------------------------------------------
-# mysql_secure_installation 5.5
-#--------------------------------------------------
-# 4. Set root password? [Y/n] Y
-# 1. Remove anonymous users? [Y/n] Y
-# 3. Disallow root login remotely? [Y/n] Y
-# 2. Remove test database and access to it? [Y/n] Y
-# 5. Reload privilege tables now? [Y/n] Y
-
-root_password = node['mysql_user']['root']['password']
-bash 'mysql_secure_installation' do
-    code <<-EOH
-        mysql -uroot <<EOF && touch /tmp/.mysql_secure_installation_complete
-DELETE FROM mysql.user WHERE User='';
-DELETE FROM mysql.user WHERE User='root' AND Host NOT IN ('localhost', '127.0.0.1', '::1');
-DROP DATABASE IF EXISTS test;
-DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%';
-SET PASSWORD FOR 'root'@'localhost' = PASSWORD('#{root_password}');
-SET PASSWORD FOR 'root'@'127.0.0.1' = PASSWORD('#{root_password}');
-SET PASSWORD FOR 'root'@'::1' = PASSWORD('#{root_password}');
-FLUSH PRIVILEGES;
-EOF
-    EOH
-    notifies :restart, 'service[mysql]'
-    only_if do
-        !File.exists?('/tmp/.mysql_secure_installation_complete')
+unless !File.exists?('/tmp/.mysql_secure_installation_complete') do
+    root_password = node['mysql_user']['root']['password']
+    #--------------------------------------------------
+    # mysql_secure_installation 5.5
+    #--------------------------------------------------
+    # 4. Set root password? [Y/n] Y
+    # 1. Remove anonymous users? [Y/n] Y
+    # 3. Disallow root login remotely? [Y/n] Y
+    # 2. Remove test database and access to it? [Y/n] Y
+    # 5. Reload privilege tables now? [Y/n] Y
+    bash 'mysql_secure_installation' do
+        code <<-EOH
+            mysql -uroot <<EOF && touch /tmp/.mysql_secure_installation_complete
+                DELETE FROM mysql.user WHERE User='';
+                DELETE FROM mysql.user WHERE User='root' AND Host NOT IN ('localhost', '127.0.0.1', '::1');
+                DROP DATABASE IF EXISTS test;
+                DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%';
+                SET PASSWORD FOR 'root'@'localhost' = PASSWORD('#{root_password}');
+                SET PASSWORD FOR 'root'@'127.0.0.1' = PASSWORD('#{root_password}');
+                SET PASSWORD FOR 'root'@'::1' = PASSWORD('#{root_password}');
+                FLUSH PRIVILEGES;
+            EOF
+        EOH
     end
+    service "mysql" do action :restart end
 end
-# service "mysql" do action :restart end
